fix: only create by stage table view for seed command
